Factor out get_mapit_municipality
