+++ pombola/south_africa/management/commands/south_africa_update_constituency_offices.py
@@ -30,10 +30,9 @@ from pombola.core.models import (OrganisationKind, Organisation, Place, PlaceKin
                          PositionTitle, Person, AlternativePersonName,
                          InformationSource)
 
-from mapit.models import Generation, Area, Code
 from ..helpers import (
-    fix_municipality_name, LocationNotFound,
-    geocode, get_na_member_lookup
+    LocationNotFound,
+    geocode, get_na_member_lookup, get_mapit_municipality
 )
 
 organisation_content_type = ContentType.objects.get_for_model(Organisation)
@@ -110,46 +109,6 @@ def verbose(message):
     if VERBOSE:
         print message
 
-
-def get_mapit_municipality(municipality, province=''):
-    municipality = fix_municipality_name(municipality)
-    mapit_current_generation = Generation.objects.current()
-
-    # If there's a municipality, try to add that as a place as well:
-    mapit_municipalities = Area.objects.filter(
-        Q(type__code='LMN') | Q(type__code='DMN'),
-        generation_high__gte=mapit_current_generation,
-        generation_low__lte=mapit_current_generation,
-        name=municipality)
-
-    mapit_municipality = None
-
-    if len(mapit_municipalities) == 1:
-        mapit_municipality = mapit_municipalities[0]
-    elif len(mapit_municipalities) == 2:
-        # This is probably a Metropolitan Municipality, which due to
-        # https://github.com/mysociety/pombola/issues/695 will match
-        # an LMN and a DMN; just pick the DMN:
-        if set(m.type.code for m in mapit_municipalities) == set(('LMN', 'DMN')):
-            mapit_municipality = [m for m in mapit_municipalities if m.type.code == 'DMN'][0]
-        else:
-            # Special cases for 'Emalahleni' and 'Naledi', which
-            # are in multiple provinces:
-            if municipality == 'Emalahleni':
-                if province=='Mpumalanga':
-                    mapit_municipality = Code.objects.get(type__code='l', code='MP312').area
-                elif province=='Eastern Cape':
-                    mapit_municipality = Code.objects.get(type__code='l', code='EC136').area
-                else:
-                    raise Exception, "Unknown Emalahleni province %s" % (province)
-            elif municipality == 'Naledi':
-                if province=='Northern Cape':
-                    mapit_municipality = Code.objects.get(type__code='l', code='NW392').area
-                else:
-                    raise Exception, "Unknown Naledi province %s" % (province)
-            else:
-                raise Exception, "Ambiguous municipality name '%s'" % (municipality,)
-    return mapit_municipality
 
 def process_office(office, commit, start_date, end_date):
     global geocode_cache, locationsnotfound, personnotfound
